I optimized the methods.
The improvement results are as follows.

|Method                          |Before Improvement (s)|After Improvement (s)|
|--------------------------------|---------------------------------|--------------------------------|
|equal_test()                    |1.27                             |0.517                           |
|not_equal_test()                |1.33                             |0.548                           |
|greater_test()                  |1.29                             |0.526                           |
|greater_or_equal_test()         |1.29                             |0.528                           |
|less_test()                     |1.29                             |0.532                           |
|less_or_equal_test()            |1.30                             |0.528                           |
|in_test()                       |1.39                             |0.624                           |
|not_in_test()                   |1.45                             |0.659                           |
|extract_test()                  |0.954                            |0.502                           |
|extract_duplicates_test()       |0.203                            |0.204                           |
|is_duplicate_test()             |264                              |0.0549                          |
|map_test()                      |0.259                            |0.247                           |
|to_dict_test()                  |0.977                            |0.512                           |
|to_dict_list_test()             |1.28                             |1.25                            |
|dicts_to_instances_test()       |1.42                             |1.29                            |
|group_by_key_test()             |1.20                             |1.20                            |
|rename_keys_test()              |15.0                             |1.16                            |
|map_for_keys_test()             |14.6                             |1.98                            |
|to_dict_with_complex_keys_test()|2.14                             |2.09                            |